cleans up gofiles
